reuse scroll bar with msgs between resume/share session view
